Complete Singly Linked List System

+++ linkedlist.py
@@ -1,6 +1,5 @@
 class Node:
     def addData(self, next=None):
-        # self.data = 0
         self.data = int(input("add: "))
         self.next = next
 
@@ -28,7 +27,28 @@ class Sll(Node):
             if valtofind == value.data:
                 self.linkedList.pop(index)
                 self.linkedList[index-1].next = self.linkedList[index]
-                self.printAllNode()
+                # self.printAllNode()
+
+    def addNewNode(self):
+        c = Node()
+        try:
+            if self.linkedList[0] == None:
+                pass
+            else:
+                c.addData()
+                placeOfNewNode = int(input("Enter the node data before the new node: "))
+                for index, value in enumerate(self.linkedList):
+                    if value.data == placeOfNewNode:
+                        self.linkedList[index].next = c
+                        c.next = self.linkedList[index+1]
+                        self.linkedList.insert(index+1, c)
+                        # self.printAllNode()
+
+        except IndexError:
+            c.addData()
+            print("linked list is empty pls enter new node")
+            self.linkedList.append(c)
+            print(f"data: {self.linkedList[-1].data} is added")
 
     def printAllNode(self):
         print("####################################################################################")
@@ -44,9 +64,22 @@ class Sll(Node):
 
 
 sll = Sll()
-sll.addNode()
-sll.addNode()
-sll.addNode()
-sll.addNode()
-sll.addNode()
-sll.deleteNode()
+while True:
+    print("###################################### SINGLY LINKED LIST #####################################################")
+    print("1. add node to linked list")
+    print("2. add new node in a middle of a linked list")
+    print("3. delete a node")
+    print("4. Show all node")
+    print("5. Exit")
+    choice = int(input("Enter Your choice: "))
+    if choice == 1:
+        sll.addNode()
+    elif choice == 2:
+        sll.addNewNode()
+    elif choice == 3:
+        sll.deleteNode()
+    elif choice == 4:
+        sll.printAllNode()
+    elif choice == 5:
+        break
+
